### 2025-07-31 #8679 @editorialbot
[PRE REVIEW]: QASM-TS 2.0: Enabling the Verification and Formalization of Hybrid Quantum-Classical Computing with OpenQASM 3.0
**Submitting author:** @comp-phys-marc (<a href="http://orcid.org/0000-0001-6591-9585">Marcus Edwards</a>)
**Repository:** https://github.com/comp-phys-marc/qasm-ts
**Branch with paper.md** (empty if default branch):
**Version:** v2.0
