### 2025-07-31 #8679 @danielskatz
👋 @comp-phys-marc - thanks for your submission.

However, JOSS cannot review it without it having an OSI-approved open source license, and it currently has no clear license.

If you want to proceed with JOSS, please add an OSI-approved license and ping me.  If not, please let me know so I can remove this submission from our system.
