### 2025-07-31 e6c8990
Update main.bib
Osaka University
