+++ paper/main.bib
@@ -198,3 +198,12 @@
 	keywords = {Computer Science - Programming Languages, Computer Science - Software Engineering},
 	file = {Preprint PDF:/Users/marcusedwards/Zotero/storage/D29EPA49/Zhao et al. - 2023 - QChecker Detecting Bugs in Quantum Programs via S.pdf:application/pdf;Snapshot:/Users/marcusedwards/Zotero/storage/QNX2SHXM/2304.html:text/html},
 }
+
+@misc{osaka_2025,
+    author={Osaka University},
+    month=march,
+    year=2025,
+    title={Oqtopus: Researchers launch open-source quantum computer operating system},
+    publisher={Phys.org},
+    url={https://phys.org/news/2025-03-oqtopus-source-quantum.html}
+}
