### 2025-07-31 6b27ae3
Update paper.md
Oqtopus
