### 2025-07-31 #8679 @comp-phys-marc
Please also note that a significant amount of development was done by myself prior to the first commit that shows up on this repo, since the package was first developed independent of the git history.
