### 2025-07-31 #8679 @comp-phys-marc
We just wanted to additionally note that Osaka University's open source quantum computer operating system project "Oqtopus" already depends on and makes use of Qasm-ts and thank the Oqtopus team for their interest in our work.
