### 2025-07-31 7feb813
Update main.bib
Properly attribute Oqtopus authors
