+++ paper/main.bib
@@ -200,10 +200,9 @@
 }
 
 @misc{osaka_2025,
-    author={Osaka University},
-    month=march,
-    year=2025,
-    title={Oqtopus: Researchers launch open-source quantum computer operating system},
-    publisher={Phys.org},
-    url={https://phys.org/news/2025-03-oqtopus-source-quantum.html}
+	author = {Aso, Noriyasu and Feluś, Karolina and Gaj, Adrian and Gokita, Shun and Góralczyk, Sławomir and Kakuko, Norihiro and Masumoto, Naoyuki and Miyaji, Kosuke and Miyanaga, Takafumi and Mori, Toshio and Noda, Kunihiro and Tsukano, Satoyuki and Ymaguchi, Masaomi and Żybort, Dobrosław},
+	month = {9},
+	title = {OQTOPUS Cloud},
+	url = {https://github.com/oqtopus-team.github.io/oqtopus-cloud},
+	year = {2024}
 }
