### 2025-07-31 3a25a83
Add countries to affiliations
