### 2025-08-01 #8679 @HectorMozo3110
@danielskatz Yes, that's right. @Eduardo111777 was my old account. This is my main account now. Apologies for the confusion, and feel free to remove the previous one. Thanks.
