### 2025-08-03 comp-phys-marc/qasm-ts#7 @HectorMozo3110
README Review — Suggestions for Improvement
👋 Hi @comp-phys-marc (Marcus Edwards),
As part of the JOSS review, I’ve gone through the `README.md` and would like to share a few suggestions to help align it more closely with the [JOSS reviewer guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). The current README is functional and includes installation and usage examples, but several important elements appear to be missing or incomplete:

---

#### ✗ 1. Missing high-level summary
There’s no introductory paragraph explaining — in simple terms for a non-specialist audience — what the software does, what its purpose is, or who might benefit from using it.

---

#### ✗ 2. Missing statement of need
The README doesn’t currently describe the motivation behind QASM-TS — what problem it addresses or how it fits into the ecosystem of OpenQASM-related tools.

---

#### ✗ 3. No community or contribution guidelines
There are no instructions for how users can contribute to the project, report issues, or request support. A short section or a link to a `CONTRIBUTING.md` file would be helpful.

---

#### ✗ 4. Minimal API documentation
While `parseFile` and `parseString` are introduced, the API documentation could benefit from additional details such as:

- argument types
- return values
- expected behavior
- error handling (if any)
- a link to comprehensive API docs, if available

---

#### ✗ 5. Broken link to OpenQASM documentation
The line that says:
> “Language documentation is provided by IBM here”

currently points to a GitHub URL that returns a 404.

* Suggested replacement:  ←
[https://openqasm.com](https://openqasm.com)

---

#### Additional suggestion — License reference
Although the `LICENSE` file is present in the repository root, consider adding a short license reference in the README for clarity. For example:

```
 © This project is licensed under the Apache License 2.0 — see the [LICENSE](LICENSE) file for details.
```

It should look like this and link to the LICENSE file in your repository root:
> © This project is licensed under the Apache License 2.0 — see the [LICENSE](LICENSE) file for details.

---

I'll continue reviewing more thoroughly over the next few days. For now, these are my initial suggestions.
If you have any questions or would like clarification on any of these points, please feel free to reach out — I'm happy to help! 🙌

(Linked to openjournals/joss-reviews#8696 as part of the ongoing JOSS review)
