### 2025-08-04 comp-phys-marc/qasm-ts#7 @comp-phys-marc
Hello @HectorMozo3110 thank you for this excellent initial set of feedback. I will get to these items over the coming week.
