### 2025-08-05 comp-phys-marc/qasm-ts#8 @HectorMozo3110
Minor Formatting & Reference Suggestions
👋 Hi @comp-phys-marc,

Thank you again for the excellent submission. The paper is well written and technically solid. I have just a couple of **minor formatting suggestions** to help polish metadata rendering and reference integrity:

---

#### 1.  Affiliation Order in Rendered PDF

In your YAML header (see [`paper/paper.md`](https://github.com/comp-phys-marc/qasm-ts/blob/master/paper/paper.md)), affiliations are currently declared as:

```yaml
authors:
  - name: Sean Kim[^1]
    affiliation: "1"
  - name: Marcus Edwards[^2]
    affiliation: "2"
affiliations:
  - name: University of British Columbia, Canada
    index: 2
  - name: George Washington University, United States of America
    index: 1
```

Although syntactically valid, this ordering causes the PDF to display UBC before GWU, even though Sean Kim is the first author. Because the renderer uses the order in the affiliations: list, it mismatches the intended author–affiliation mapping.
**Suggested fix:** Swap the affiliation entries to match author order:

```
affiliations:
  - name: George Washington University, United States of America
    index: 1
  - name: University of British Columbia, Canada
    index: 2
```
For consistency and compliance, you might refer to the JOSS metadata example in the official guide:
https://joss.readthedocs.io/en/latest/example_paper.html

---

### 2.  Broken or Incomplete Reference Links in `main.bib`

I’ve reviewed the bibliography ([`paper/main.bib`](https://github.com/comp-phys-marc/qasm-ts/blob/master/paper/main.bib)) and identified two entries that may benefit from updates or clarification:

---

✗  **Broken URL**

**Line 206**
**Oqtopus Cloud**
https://github.com/oqtopus-team.github.io/oqtopus-cloud → Currently returns a **404 Not Found**.
 BibTeX entry location: [Line 202](https://github.com/comp-phys-marc/qasm-ts/blob/master/paper/main.bib#L202)

**Suggested replacement**:
→ [`https://github.com/oqtopus-team/oqtopus-cloud`](https://github.com/oqtopus-team/oqtopus-cloud)

---

✗  **Broken URL**

**Line 91**
**QEC Decoder Toolkit – Quantinuum Docs**
→ The URL in the `url` and `note` fields appears to be malformed or outdated:
[`https://docs.quantinuum.com/h-series/trainings/getting_started/pytket_quantinuum/qec_decoder_toolkit/...`](https://docs.quantinuum.com/h-series/trainings/getting_started/pytket_quantinuum/qec_decoder_toolkit/Quantinuum_hseries_qec_decoder_toolkit.html)

BibTeX entry location: [Line 89](https://github.com/comp-phys-marc/qasm-ts/blob/master/paper/main.bib#L89)

---
> **Suggestion**: For each of the above, verifying if a working DOI or updated link is available would strengthen citation stability and ensure long-term reproducibility.

(Linked to https://github.com/openjournals/joss-reviews/issues/8696 as part of the ongoing JOSS review)
