### 2025-08-05 comp-phys-marc/qasm-ts#9 @HectorMozo3110
Observations after Local Testing
@comp-phys-marc,

###  Lexer Testing Report

I ran several tests on the parser, including the unit tests described in the `README.md`. The system behaves as expected and correctly generates the abstract syntax tree (AST) in all documented scenarios.

However, multiple lexer tests fail due to token count mismatches. For instance:

- `Expected 148 to equal 147`
- `Expected 67 to equal 68`
- `Expected 1071 to equal 1070`

These mismatches suggest that the expected token arrays used in the tests (see [`spec/lexer.spec.ts`](https://github.com/comp-phys-marc/qasm-ts/blob/master/spec/lexer.spec.ts)) may be outdated and no longer reflect the current behavior of the lexer. Despite these discrepancies, the generated AST structure is valid and consistently correct.

To confirm this, I manually inspected the token outputs and observed consistent count differences in the following QASM files located in [`spec/qasm/generic`](https://github.com/comp-phys-marc/qasm-ts/tree/master/spec/tokens/generic):

- **ipe.qasm:** Expected 115, Got 116

- **msd.qasm:** Expected 1070, Got 1071

- **rus.qasm:** Expected 147, Got 148

- **varteleport.qasm:** Expected 205, Got 206

Additionally, although I cannot confirm the exact cause, I suspect that these lexer test mismatches may be related to platform-specific differences such as newline handling (\r\n vs \n) or UTF-8 file encoding (e.g., presence of BOM) on Windows systems. On a clean Linux environment, all lexer tests pass without issues. It might be helpful to include a short note in the README under the testing section to inform users that such variations could cause test failures in certain environments, even if the lexer itself is functioning correctly. If aligning the token expectations across platforms is not feasible, simply documenting the discrepancy should be sufficient to avoid confusion for future users and contributors.

**I will continue monitoring the repository and re-running tests as updates are made.**

> **Conclusion**
> The software meets the expectations for this section, with only minor adjustments needed to align the lexer test expectations with the current output.

Any future updates will be posted in this issue thread.

Have a great week, and I’ll be in touch over the coming weeks.
Feel free to reach out if you have any questions or need clarification.

(Linked to https://github.com/openjournals/joss-reviews/issues/8696 as part of the ongoing JOSS review)
