### 2025-08-05 #8696 @danielskatz
👋 @apizzuto - when you are ready to get started, you need to use the command `@editorialbot generate my checklist` to create your review checklist. @editorialbot commands need to be the first thing in a new comment.
