### 2025-08-05 comp-phys-marc/qasm-ts#9 @comp-phys-marc
Thank you very much for bringing these discrepancies to our attention. I will investigate them and see exactly what is happening in each case.
