### 2025-08-05 comp-phys-marc/qasm-ts#9 @comp-phys-marc
Hello, I am a bit puzzled. I stepped through each of the lexer tests, looking specifically for the cases you highlighted, and was not able to reproduce these discrepancies. As you can see in the attached screenshot, when I inspect the same assertions on the token lengths that you mention, the assertions all pass. I wonder what is making the difference between what you see and what I am seeing.
