### 2025-08-05 comp-phys-marc/qasm-ts#8 @comp-phys-marc
Hello, thanks again for your thoughtful feedback. I have created a PR to address these things [here](https://github.com/comp-phys-marc/qasm-ts/pull/10). Please let me know if these changes look good!
