### 2025-08-06 comp-phys-marc/qasm-ts#9 @HectorMozo3110
Hi! I tested the project on both **Linux** and **Windows**, and encountered minor `token count mismatch` errors when running `npm test` on Windows. These seem to stem from newline inconsistencies (`\r\n` vs `\n`) or slightly outdated token expectations.

To help clarify and improve cross-platform test reproducibility, I opened a pull request:
 **[#11 – Fix: Avoid token count mismatch failure in lexer tests on Windows](https://github.com/comp-phys-marc/qasm-ts/pull/11)**

The PR includes:
- Normalization of line endings (`\r\n`, `\r`) to `\n` to avoid false mismatches
- Detailed console error output when tokens differ (e.g., missing, extra, or misaligned tokens)
- No change to core parsing or logic — AST output remains correct and validated

Attached is a screenshot of the token mismatch errors I encountered on Windows:

Let me know if you'd prefer a cleaner patch or stricter token checking.
