### 2025-08-07 05e9616
Argument types
