### 2025-08-07 02a80a2
Return types
