+++ readme.md
@@ -34,6 +34,8 @@ let ast = parseFile("<file-path>");
 let ast = parseString("<qasm-string>");
 ```
 
+The return type for both `parseFile` and `parseString` is `Array<AstNode>`, unless the `stringify` parameter is `true`, in which case the return is a `String`.
+
 ## Example I/O
 
 ### Input: `alignment.qasm` ([source](https://github.com/openqasm/openqasm/blob/main/examples/alignment.qasm))
