+++ readme.md
@@ -40,6 +40,8 @@ The parser is able to recognize and handle 19 distinct types of syntax errors, w
 exported in `errors.ts`. While this is not an advanced semantic or static analysis, it should enable 
 users to basically validate their OpenQASM 2.0 or 3.0 code.
 
+Conprehensive API docs can be found in the `docs/` directory.
+
 ## Example I/O
 
 ### Input: `alignment.qasm` ([source](https://github.com/openqasm/openqasm/blob/main/examples/alignment.qasm))
