### 2025-08-07 comp-phys-marc/qasm-ts#7 @comp-phys-marc
Making the changes request in this ongoing WIP PR: https://github.com/comp-phys-marc/qasm-ts/pull/12
