### 2025-08-07 9dcc00b
Summary and motivation
