### 2025-08-07 c12cc33
Contributing guidelines
