### 2025-08-08 comp-phys-marc/qasm-ts#7 @HectorMozo3110
Looks good to me.
