### 2025-08-08 comp-phys-marc/qasm-ts#9 @HectorMozo3110
👋 Hi, @comp-phys-marc  I’ll keep this open in case I notice anything else, and will close it at the end if I don’t. Regarding the tests, I’ll continue running more while we wait for the evaluation from the other reviewers.
