### 2025-08-12 #8696 @xuanxu
> 👋 openjournals/dev - I don't think the reference check command has worked for this submission in the pre-review issue or here. Can you see if there is some error being generated that isn't being pushed back to us here?

There is a formatting error in the paper's metadata. [This PR](https://github.com/comp-phys-marc/qasm-ts/pull/13) should fix it.
