### 2025-08-12 comp-phys-marc/qasm-ts#9 @HectorMozo3110
Hi, I’ll be closing this issue. In PR #14 , I only updated the commit author and email for proper record keeping.
I’ve also finished running all local tests, and everything is working correctly. ✅
