### 2025-08-13 #8696 @comp-phys-marc
Yes, that DOI appears to be correct.
