### 2025-08-25 comp-phys-marc/qasm-ts#16 @apizzuto
Request for more thorough example usage in docs
Part of https://github.com/openjournals/joss-reviews/issues/8696

I was wondering if you could expand in the "Example I/O" section of the docs on how exactly to obtain the output you show? I think this is straightforward enough for people familiar with typescript, but I think in installation docs you want to assume as little knowledge for the user as possible.

e.g. Change "Run with `const . . . `" to

Run the following commands:
```
foo
bar
```
