### 2025-08-25 comp-phys-marc/qasm-ts#17 @apizzuto
Small comments as part of JOSS review
Part of https://github.com/openjournals/joss-reviews/issues/8696

I have two EXTREMELY minor nits (small enough that I just made one issue because it felt like overkill to make two separate ones):

(1) You have some performance benchmarks in the paper. Is there a way to reproduce these benchmarks (at least for the QASM-TS part, if not for everything)? Typically when there are performance claims in a JOSS submission, it's nice to have them confirmable. If this is not possible, I think that's okay so long as we document here why it's not

(2) Super minor typos in the README that I thought I'd point out:
"compatability" -> "compatibility"
"conprehensive" -> "comprehensive"
