### 2025-08-25 #8696 @apizzuto
Apologies for the delay in the review.

Overall I found the paper to be well written, the code to be straightforward, and the API docs were quite nice. I left two issues on the GitHub repo (linked above), but ultimately these represent minor changes.

Pending those issues being fixed, I will check off the last few items on my checklist, and then this will have the green light from my end.

Thanks for a great submission!
