### 2025-08-25 comp-phys-marc/qasm-ts#17 @comp-phys-marc
Thanks @apizzuto  for your review and your great feedback! I have created a PR to address the minor typos you found. Also, to run the benchmarks and reproduce our results, please see this [repo](https://github.com/seankim658/qasm-parser-testing) and follow the steps outlined in the README.

Let me know if you have any questions!
