### 2025-08-31 comp-phys-marc/qasm-ts#20 @Abinashbunty
Modify documentation structure for more efficient guide for new users
Hi authors! I created this issue as part of JOSS review here: https://github.com/openjournals/joss-reviews/issues/8696

I just had couple of suggestions about the structure of the documentation.

- Navigating from index.html, I feel new users of the library might not find the relevant information about the functions or enough examples to guide them in using it. Maybe some more examples will help, at least of the major functions.
- The top-down walkthrough of the TOC can be updated a bit so that it is well understood by users how to follow the docs. Since there is no "Next page" functionality of the docs, I created this issue just for some restructure.

If time permits, you can follow the [Diataxis](https://diataxis.fr/) approach. Otherwise, just a minor restructure is also fine.
