### 2025-09-04 comp-phys-marc/qasm-ts#16 @comp-phys-marc
Hello, I have made a pull request (above) to address your comments. Please let me know your thoughts!
