### 2025-09-05 835b972
Add better module, function, and class level documentation
