### 2025-09-08 #8696 @Abinashbunty
Thank you @comp-phys-marc for the update. I went through the PR and the generated docs. It is great! Kudos to @seankim658 for the changes. 🚀

@danielskatz My review is complete and all items in the checklist have been marked as complete.
