### 2025-09-08 #8696 @danielskatz
👋 @apizzuto - can you also confirm that you are ready to have this be accepted?
