### 2025-09-13 #8696 @danielskatz
@comp-phys-marc check references
