### 2025-09-13 #8696 @danielskatz
@comp-phys-marc - I've suggested a number of small changes in https://github.com/comp-phys-marc/qasm-ts/pull/27. Please merge this, or let me know what you disagree with.

Also, I'm not sure that this sentence makes sense as written "By our typed implementation of an OpenQASM 3.0 parser, we implement a system that infers types from QASM syntax." but I'm not sure how best to change it.  Maybe "Our typed OpenQASM 3.0 parser implements a system that infers types from QASM syntax."?
