### 2025-09-14 e6fc8ee
Update paper.md
Adopt suggestion.
