### 2025-09-14 #8696 @danielskatz
@comp-phys-marc generate pdf
