### 2025-09-17 #8696 @danielskatz
This is the correct style for JOSS in this case
